--- a/src/chess/bitboard.cc
+++ b/src/chess/bitboard.cc
@@ -28,7 +28,7 @@
 #include "chess/bitboard.h"
 #include "utils/exception.h"
 
-#include <iostream>
+#include <vector>
 
 namespace lczero {
 
@@ -318,6 +318,12 @@ static const BitBoard kBishopAttacks[] = {
     0x0028448201000000ULL, 0x0050880402010000ULL, 0x00A0100804020100ULL,
     0x0040201008040201ULL};
 
+static const std::pair<int, int> kRookDirections[] = {
+    {1, 0}, {-1, 0}, {0, 1}, {0, -1}};
+
+static const std::pair<int, int> kBishopDirections[] = {
+    {1, 1}, {-1, 1}, {1, -1}, {-1, -1}};
+
 std::vector<unsigned short> BuildMoveIndices() {
   std::vector<unsigned short> res(4 * 64 * 64);
   for (size_t i = 0; i < sizeof(kIdxToMove) / sizeof(kIdxToMove[0]); ++i) {
@@ -336,30 +342,180 @@ const int kQueenCastleIndex =
 static MagicBitBoards magic_bitboards;
 }  // namespace
 
-const BitBoard MagicBitBoards::kRookMagicNumbers[64];
-const BitBoard MagicBitBoards::kBishopMagicNumbers[64];
+const BitBoard MagicBitBoards::kRookMagicNumbers[64] = {
+    0x088000102088C001ULL, 0x10C0200040001000ULL, 0x83001041000B2000ULL,
+    0x0680280080041000ULL, 0x488004000A080080ULL, 0x0100180400010002ULL,
+    0x040001C401021008ULL, 0x02000C04A980C302ULL, 0x0000800040082084ULL,
+    0x5020C00820025000ULL, 0x0001002001044012ULL, 0x0402001020400A00ULL,
+    0x00C0800800040080ULL, 0x4028800200040080ULL, 0x00A0804200802500ULL,
+    0x8004800040802100ULL, 0x0080004000200040ULL, 0x1082810020400100ULL,
+    0x0020004010080040ULL, 0x2004818010042800ULL, 0x0601010008005004ULL,
+    0x4600808002001400ULL, 0x0010040009180210ULL, 0x020412000406C091ULL,
+    0x040084228000C000ULL, 0x8000810100204000ULL, 0x0084110100402000ULL,
+    0x0046001A00204210ULL, 0x2001040080080081ULL, 0x0144020080800400ULL,
+    0x0840108400080229ULL, 0x0480308A0000410CULL, 0x0460324002800081ULL,
+    0x620080A001804000ULL, 0x2800802000801006ULL, 0x0002809000800800ULL,
+    0x4C09040080802800ULL, 0x4808800C00800200ULL, 0x0200311004001802ULL,
+    0x0400008402002141ULL, 0x0410800140008020ULL, 0x000080C001050020ULL,
+    0x004080204A020010ULL, 0x0224201001010038ULL, 0x0109001108010004ULL,
+    0x0282004844020010ULL, 0x8228180110040082ULL, 0x0001000080C10002ULL,
+    0x024000C120801080ULL, 0x0001406481060200ULL, 0x0101243200418600ULL,
+    0x0108800800100080ULL, 0x4022080100100D00ULL, 0x0000843040600801ULL,
+    0x8301000200CC0500ULL, 0x1000004500840200ULL, 0x1100104100800069ULL,
+    0x2001008440001021ULL, 0x2002008830204082ULL, 0x0010145000082101ULL,
+    0x01A2001004200842ULL, 0x1007000608040041ULL, 0x000A08100203028CULL,
+    0x02D4048040290402ULL};
+const BitBoard MagicBitBoards::kBishopMagicNumbers[64] = {
+    0x0008201802242020ULL, 0x0021040424806220ULL, 0x4006360602013080ULL,
+    0x0004410020408002ULL, 0x2102021009001140ULL, 0x08C2021004000001ULL,
+    0x6001031120200820ULL, 0x1018310402201410ULL, 0x401CE00210820484ULL,
+    0x001029D001004100ULL, 0x2C00101080810032ULL, 0x0000082581000010ULL,
+    0x10000A0210110020ULL, 0x200002016C202000ULL, 0x0201018821901000ULL,
+    0x006A0300420A2100ULL, 0x0010014005450400ULL, 0x1008C12008028280ULL,
+    0x00010010004A0040ULL, 0x3000820802044020ULL, 0x0000800405A02820ULL,
+    0x8042004300420240ULL, 0x10060801210D2000ULL, 0x0210840500511061ULL,
+    0x0008142118509020ULL, 0x0021109460040104ULL, 0x00A1480090019030ULL,
+    0x0102008808008020ULL, 0x884084000880E001ULL, 0x040041020A030100ULL,
+    0x3000810104110805ULL, 0x04040A2006808440ULL, 0x0044040404C01100ULL,
+    0x4122B80800245004ULL, 0x0044020502380046ULL, 0x0100400888020200ULL,
+    0x01C0002060020080ULL, 0x4008811100021001ULL, 0x8208450441040609ULL,
+    0x0408004900008088ULL, 0x0294212051220882ULL, 0x000041080810E062ULL,
+    0x10480A018E005000ULL, 0x80400A0204201600ULL, 0x2800200204100682ULL,
+    0x0020200400204441ULL, 0x0A500600A5002400ULL, 0x801602004A010100ULL,
+    0x0801841008040880ULL, 0x10010880C4200028ULL, 0x0400004424040000ULL,
+    0x0401000142022100ULL, 0x00A00010020A0002ULL, 0x1010400204010810ULL,
+    0x0829910400840000ULL, 0x0004235204010080ULL, 0x1002008143082000ULL,
+    0x11840044440C2080ULL, 0x2802A02104030440ULL, 0x6100000900840401ULL,
+    0x1C20A15A90420200ULL, 0x0088414004480280ULL, 0x0000204242881100ULL,
+    0x0240080802809010ULL};
 
-MagicBitBoards::MagicParams MagicBitBoards::kRookMagicParams[64];
-MagicBitBoards::MagicParams MagicBitBoards::kBishopMagicParams[64];
+MagicBitBoards::MagicParams MagicBitBoards::rook_magic_params_[64];
+MagicBitBoards::MagicParams MagicBitBoards::bishop_magic_params_[64];
+
+BitBoard MagicBitBoards::rook_attacks_table_[102400];
+BitBoard MagicBitBoards::bishop_attacks_table_[5248];
 
 MagicBitBoards::MagicBitBoards() {
-  // Initialize for all board squares.
-  for (int square = 0; square < 64; square++) {
+  // Initialize masks for all board squares.
+  for (unsigned square = 0; square < 64; square++) {
     const BoardSquare b_sq(square);
 
-    // Calculate occupancy masks by subtracting the board edges from the total
-    // attack bitboards.
-    kRookMagicParams[square].mask_ =
+    // Calculate relevant occupancy masks by subtracting the board edges from
+    // the total attack bitboards.
+    rook_magic_params_[square].mask_ =
         kRookAttacks[square] - BoardSquare(b_sq.row(), 0) -
         BoardSquare(b_sq.row(), 7) - BoardSquare(0, b_sq.col()) -
         BoardSquare(7, b_sq.col());
-    kBishopMagicParams[square].mask_ =
+    bishop_magic_params_[square].mask_ =
         kBishopAttacks[square] - BitBoard(0xFF818181818181FFULL);
-
-    // Print the table index size.
-    std::cout << "Number of relevant occupancy bits = "
-              << kBishopMagicParams[square].mask_.count() << std::endl;
   }
+
+  // Build attack tables.
+  BuildAttackTable(kRookMagicNumbers, rook_magic_params_, rook_attacks_table_,
+                   kRookDirections);
+  BuildAttackTable(kBishopMagicNumbers, bishop_magic_params_,
+                   bishop_attacks_table_, kBishopDirections);
+}
+
+void MagicBitBoards::BuildAttackTable(const BitBoard* magic_numbers,
+                                      MagicParams* magic_params,
+                                      BitBoard* attacks_table,
+                                      const std::pair<int, int>* directions) {
+  // Offset into lookup table.
+  uint32_t table_offset = 0;
+
+  // Initialize for all board squares.
+  for (unsigned square = 0; square < 64; square++) {
+    const BoardSquare b_sq(square);
+
+    // Cache the relevant occupancy board squares.
+    std::vector<BoardSquare> occupancy_squares;
+
+    for (auto temp_sq : magic_params[square].mask_) {
+      occupancy_squares.emplace_back(temp_sq);
+    }
+
+    // Set the number of shifted bits.
+    magic_params[square].shift_bits_ = 64 - occupancy_squares.size();
+
+    // Set the lookup table offset.
+    magic_params[square].table_offset_ = table_offset;
+
+    // Clear the attack table.
+    for (int i = 0; i < (1 << occupancy_squares.size()); i++) {
+      attacks_table[table_offset + i] = 0;
+    }
+
+    // Build square attack table for every possible masked occupancy bitboard.
+    for (int i = 0; i < (1 << occupancy_squares.size()); i++) {
+      BitBoard occupancy(0);
+
+      for (size_t bit = 0; bit < occupancy_squares.size(); bit++) {
+        occupancy.set_if(occupancy_squares[bit], (1 << bit) & i);
+      }
+
+      // Calculate attack bitboard corresponding to this occupancy bitboard.
+      BitBoard attack(0);
+
+      for (int j = 0; j < 4; j++) {
+        auto direction = directions[j];
+        auto dst_row = b_sq.row();
+        auto dst_col = b_sq.col();
+        while (true) {
+          dst_row += direction.first;
+          dst_col += direction.second;
+          if (!BoardSquare::IsValid(dst_row, dst_col)) break;
+          const BoardSquare destination(dst_row, dst_col);
+          attack.set(destination);
+          if (occupancy.get(destination)) break;
+        }
+      }
+
+      // Calculate magic index.
+      uint64_t index = occupancy.as_int();
+      index *= magic_numbers[square].as_int();
+      index >>= magic_params[square].shift_bits_;
+
+      // Update the table.
+      if (attacks_table[table_offset + index] != 0 &&
+          attacks_table[table_offset + index] != attack) {
+        // Sanity check.
+        throw Exception("Invalid magic number!");
+      }
+
+      attacks_table[table_offset + index] = attack;
+    }
+
+    // Update table offset.
+    table_offset += (1 << occupancy_squares.size());
+  }
+}
+
+inline BitBoard MagicBitBoards::GetRookAttacks(const BoardSquare rook_square,
+                                               const BitBoard pieces) const {
+  // Calculate magic index.
+  uint8_t square = rook_square.as_int();
+
+  uint64_t index = (pieces * rook_magic_params_[square].mask_).as_int();
+  index *= kRookMagicNumbers[square].as_int();
+  index >>= rook_magic_params_[square].shift_bits_;
+
+  // Return attack bitboard.
+  return rook_attacks_table_[rook_magic_params_[square].table_offset_ + index];
+}
+
+inline BitBoard MagicBitBoards::GetBishopAttacks(
+    const BoardSquare bishop_square, const BitBoard pieces) const {
+  // Calculate magic index.
+  uint8_t square = bishop_square.as_int();
+
+  uint64_t index = (pieces * bishop_magic_params_[square].mask_).as_int();
+  index *= kBishopMagicNumbers[square].as_int();
+  index >>= bishop_magic_params_[square].shift_bits_;
+
+  // Return attack bitboard.
+  return bishop_attacks_table_[bishop_magic_params_[square].table_offset_ +
+                               index];
 }
 
 Move::Move(const std::string& str, bool black) {