… src/chess/board.cc