This PR adds "fancy magic bitboards" to move generation.

The following measurements all share the command line `lc0 benchmark --backend=random --smart-pruning-factor=0 --nodes=1000000 --nncache=1000000`. A 95% confidence interval for the average of the last reported nps value over 25 runs is given for the starting position, the opening position https://lichess.org/editor/rn2kb1r/pp3ppp/2p1pnb1/8/3P1N1q/4B1N1/PPPQ1PPP/R3KB1R_b_KQkq_-_5_9 and the endgame position https://lichess.org/editor/8/2R2P2/1p4K1/6R1/1r6/5rP1/1k6/2q5_w_-_-_0_55 .

The measurements are performed on an Intel Core i5-5300U CPU @ 2.30GHz with 4 cores and L1d cache: 32K, L1i cache: 32K, L2 cache: 256K, L3 cache: 3072K.

For 1 search thread there is a substantial speedup for all positions:

position  |  master `0af66eee` |  #628  | speedup
------ | ----- | ----------- | -----------
start | 186968 ± 313 | 189317 ± 229 | 1.3%
opening | 141347 ± 340 | 144938 ± 354 | 2.5%
endgame | 154753 ± 248 | 162247 ± 208 | 4.8%

For 2 search threads the performance drops slightly on my system for all positions:

position  |  master `0af66eee` |  #628  | speedup
------ | ----- | ----------- | -----------
start | 227397 ± 565 | 226473 ± 745 | -0.4%
opening | 180583 ± 442 | 178087 ± 569 | -1.4%
endgame | 191124 ± 530 | 188983 ± 463 | -1.1%

I think this can be completely attributed to the fact that there are more cache misses as the 3MB L3 cache is shared between all cores on my CPU (and not that big really). Note that the rook attack tables are ~800KB and the bishop attack tables are ~40KB.

The speedup will probably be positive for 2 search threads on higher-end CPUs. Feel free to test!