@mooskagh 2018-12-29
FYI there are pure movegen tests (without any mcts) in board_test.cc.
They don't measure time, but that can be added. (also can be measured by external program, like bash `time`)