@gsobala 2018-12-30
As there have been recent search improvements, I have merged this with current master (6a639b6e9f7d2aaec0a81de6f8c0bd683461b756) and tested it against current master with 2 threads and starting position on an 8-core Xeon with L1 cache 32k/32kx8 L2 cache 8MB and L3 cache 11MB on Ubuntu, with your opening position fen and ending fen. I took the mean of 20 tests. 

PR628 was 1.25% faster than master on the opening FEN and 2.3% faster than master on the ending FEN.