@ddobbelaere 2018-12-30
The chessboard test is 25% faster with this PR.

Thanks for the tests @gsobala. I obtain more or less the same speedup results with current master.

I should also mention the single-run measurements of sv in discord (on his 2x Intel Xeon silver 4108, 11M L3 cache each @1.8 GHz).

```
master nps
1 thread: 112908
2 threads: 129790
4 threads: 125918
8 threads: 109047
16 threads: 103957
32 threads: 95132

PR628 nps
1 thread: 112711
2 threads: 116351
4 threads: 129308
8 threads: 98344.8
16 threads: 90542.6
32 threads: 82001.5
```
Before moving on, I'd like to investigate the exact cause of the slowdown in some cases in my machine by trying to pinpoint the root cause (cache misses?) and add possible optimizations.