@ddobbelaere 2018-12-30
I verified that cache misses are indeed the cause of the slowdown in some cases.

Tried a lot of code optimizations to further minimize CPU cycles (e.g. by trying faster PEXT instructions present on modern CPUs to index the attacks tables, effectively eliminating index calculation with magic numbers), to no avail, as this doesn't solve cache misses.

I think that `lc0` really differs from A/B engines in this respect. In A/B engines, nps are orders of magnitudes higher (~several Mnps) and cache misses are reduced by the "batched movegen" behavior. `lc0` does a lot of other stuff in between generating positions (MCTS, node cache lookup...) that fight for the same cache.

This explains why `chessboard_test` runs 25% faster using magic bitboard (even 27% with PEXT instruction), but using magic bitboards in `lc0` itself hasn't the hoped effect due to more costly cache misses.