@ddobbelaere 2019-01-01
Turns out that with LTO enabled, I experience no slowdown anymore (speedup for all considered cases). Magic bitboards are implemented in #640, as an additional layer on top of #638 to avoid merge conflicts with this PR.